add bloom support
