+++ Makefile
@@ -108,9 +108,3 @@ docs:             ## Build the documentation.
 # 	@mv setup.py .github/backup
 # 	@echo "You have switched to https://python-poetry.org/ package manager."
 # 	@echo "Please run 'poetry shell' or 'poetry run developergpt'"
-
-
-# This project has been generated from rochacbruno/python-project-template
-# __author__ = 'rochacbruno'
-# __repo__ = https://github.com/rochacbruno/python-project-template
-# __sponsor__ = https://github.com/sponsors/rochacbruno/
