+++ requirements.txt
@@ -4,3 +4,4 @@ tiktoken
 rich
 inquirer
 prompt_toolkit
+text_generation
